django-portpholio/portpholio/requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==38.0.3
distlib==0.3.7
Django==4.0.6
django-bootstrap3==23.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
ez-setup==0.9
Faker==13.15.0
filelock==3.12.2
Flask==2.3.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
misaka==2.1.1
numpy==1.23.1
packaging==23.1
paramiko==2.12.0
Pillow==9.2.0
platformdirs==3.9.1
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlparse==0.4.2
sudo==1.0.0
typing_extensions==4.7.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.24.2
Werkzeug==2.3.6
